Moved PostInfo & Geographical related values to subdir
