+++ src/Value/FullAddress.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Flanders\BasicRegisters\Value;
 
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\AbstractGeographicalNames;
+
 /**
  * Collection of full address geographical names.
  */
