+++ src/Value/Post/PostInfoId.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Flanders\BasicRegisters\Value\Post;
+
+use DigipolisGent\Flanders\BasicRegisters\Value\AbstractId;
+
+/**
+ * Post info object ID.
+ */
+final class PostInfoId extends AbstractId
+{
+
+}
