--- src/Value/PostInfoId.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DigipolisGent\Flanders\BasicRegisters\Value;
-
-/**
- * Post info object ID.
- */
-final class PostInfoId extends AbstractId
-{
-
-}
