+++ src/Value/Street/StreetName.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Flanders\BasicRegisters\Value\Street;
 
-use DigipolisGent\Flanders\BasicRegisters\Value\AbstractWithGeographicalName;
-use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\AbstractWithGeographicalName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
 use DigipolisGent\Value\ValueInterface;
 
 /**
@@ -24,7 +24,7 @@ final class StreetName extends AbstractWithGeographicalName
      * Create a new street name.
      *
      * @param \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId $streetNameNameId
-     * @param \DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName $geographicalName
+     * @param \DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName $geographicalName
      */
     public function __construct(StreetNameId $streetNameNameId, GeographicalName $geographicalName)
     {
