+++ tests/Normalize/FromJson/StreetNameNormalizerTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace DigipolisGent\Tests\Flanders\BasicRegisters\Normalize\FromJson;
 
 use DigipolisGent\Flanders\BasicRegisters\Normalizer\FromJson\StreetNameNormalizer;
-use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetName;
 use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId;
