+++ tests/Value/Geographical/GeographicalNamesTest.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value;
+namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value\Geographical;
 
-use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
-use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalNames;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \DigipolisGent\Flanders\BasicRegisters\Value\AbstractGeographicalNames
- * @covers \DigipolisGent\Flanders\BasicRegisters\Value\GeographicalNames
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Value\Geographical\AbstractGeographicalNames
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalNames
  */
 class GeographicalNamesTest extends TestCase
 {
