+++ tests/Value/Post/PostInfoTest.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value;
+namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value\Post;
 
-use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
-use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalNames;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
-use DigipolisGent\Flanders\BasicRegisters\Value\PostInfo;
-use DigipolisGent\Flanders\BasicRegisters\Value\PostInfoId;
+use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfo;
+use DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfoId;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \DigipolisGent\Flanders\BasicRegisters\Value\PostInfo
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Value\Post\PostInfo
  */
 class PostInfoTest extends TestCase
 {
