+++ tests/Value/Street/StreetNameDetailTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace DigipolisGent\Tests\Flanders\BasicRegisters\Value\Street;
 
-use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalName;
-use DigipolisGent\Flanders\BasicRegisters\Value\GeographicalNames;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalName;
+use DigipolisGent\Flanders\BasicRegisters\Value\Geographical\GeographicalNames;
 use DigipolisGent\Flanders\BasicRegisters\Value\LanguageCode;
 use DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityName;
 use DigipolisGent\Flanders\BasicRegisters\Value\Locality\LocalityNameId;
@@ -14,7 +14,7 @@ use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameId;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \DigipolisGent\Flanders\BasicRegisters\Value\AbstractWithGeographicalNames
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Value\Geographical\AbstractWithGeographicalNames
  * @covers \DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNameDetail
  */
 class StreetNameDetailTest extends TestCase
